# WKB reinterpret cast marshaling hack